added benchmark for hoisting & fixed caching
use path in grammar as key for caching elements; using the elements
itself doesn't work because the reference will be different for the
content assist grammar run.
this does not work for generated elements; the only place that generated
elements are necessary is when dealing with nested identical paths, in
this case caching is disabled.

the benchmark test case does not validate the result since we are only
interested in the runtime.