fix for endless prefix problem
our algorithm can't deal with recursive prefixes. that's not a huge
problem though since ANTLR uses LL(*) which can't deal with these
constructs either.

solution for the moment: if the minimal path difference analysis errors
out check if more than 2 path are recursive. if so, throw an exception

possible improvement for later: just consider predicates and not tokens
- similar to ANTLR which falls back to LL(1)