rewrite of minimal path difference analysis
old version was too slow to recognize nested identical paths because the
complete context needed to be analyzed.

now the token analysis saves meta information about whether a given
token is from the context or the target element. using this we can
detect nested identical paths by checking if the context starts at the
same position if two token lists are identical.
this can be done when ever the index list is full (no gaps), which is a
lot sooner than the end of the tokenCombinations call.